Fix some issues with groups
Now by default, it is recommended that `set_group` simply returns the
parent group unless it specifically overrides some behaviour that should
be carried through to children.

From this point, consider the return value of `set_group` to be the
parent group of children, not just the group of the current node. If the
current node's group is passed, then `behind_parent` will no longer work
due to the way that groups are ordered.